rename(#63): alert → notification 페이지명 변경
