+++ src/components/alert/AlertWindow.tsx
@@ -2,7 +2,7 @@ import styled from "styled-components";
 import { theme } from "@/styles/theme";
 import Image from "next/image";
 import { useCallback, useEffect, useRef, useState } from "react";
-import AlertBox from "../mypage/alert/AlertBox";
+import AlertBox from "../mypage/notification/AlertBox";
 import { EX_ALARM, EX_REQUEST_ALARM } from "@/data/mypage";
 import { useRouter } from "next/navigation";
 
@@ -20,14 +20,17 @@ const AlertWindow = (
 
   const alertWindowRef = useRef<HTMLDivElement>(null);
 
-  const handleClickOutside = useCallback((event: MouseEvent) => {
-    if (
-      alertWindowRef.current &&
-      !alertWindowRef.current.contains(event.target as Node)
-    ) {
-      onClose();
-    }
-  }, [onClose]);
+  const handleClickOutside = useCallback(
+    (event: MouseEvent) => {
+      if (
+        alertWindowRef.current &&
+        !alertWindowRef.current.contains(event.target as Node)
+      ) {
+        onClose();
+      }
+    },
+    [onClose]
+  );
 
   useEffect(() => {
     document.addEventListener("mousedown", handleClickOutside);
@@ -75,7 +78,7 @@ const AlertWindow = (
                 <AlertBox
                   key={index}
                   content={data.content}
-                  time={data.time}
+                  createdAt={data.time}
                   read={data.read}
                   size="small"
                 />
@@ -88,7 +91,7 @@ const AlertWindow = (
                 <AlertBox
                   key={index}
                   content={data.content}
-                  time={data.time}
+                  createdAt={data.time}
                   read={data.read}
                   size="small"
                 />
