+++ src/data/tab.ts
@@ -9,7 +9,7 @@ export const HEADER_MODAL_TAB = [
 
 export const MYPAGE_TAB = [
     { id: 1, menu: '내 정보', url: "/mypage/profile" },
-    { id: 2, menu: '알림', url: "/mypage/alert" },
+    { id: 2, menu: '알림', url: "/mypage/notification" },
     { id: 3, menu: '내가 작성한 글', url: "/mypage/post" },
     { id: 4, menu: '내 평가', url: "/mypage/review" },
     { id: 5, menu: '차단목록', url: "/mypage/blocked" },
